IOS-5033 Extracted models
