+++ Tangem/Modules/Send/Summary/SendAmountSummaryView.swift
@@ -8,35 +8,6 @@
 
 import SwiftUI
 
-class SendAmountSummaryViewModel: Identifiable {
-    let amount: String
-    let amountFiat: String
-    let tokenIconName: String
-    let tokenIconURL: URL?
-    let tokenIconCustomTokenColor: Color?
-    let tokenIconBlockchainIconName: String?
-    let isCustomToken: Bool
-
-    init(
-        amount: String,
-        amountFiat: String,
-        tokenIconName: String,
-        tokenIconURL: URL?,
-        tokenIconCustomTokenColor: Color?,
-        tokenIconBlockchainIconName: String?,
-        isCustomToken: Bool
-
-    ) {
-        self.amount = amount
-        self.amountFiat = amountFiat
-        self.tokenIconName = tokenIconName
-        self.tokenIconURL = tokenIconURL
-        self.tokenIconCustomTokenColor = tokenIconCustomTokenColor
-        self.tokenIconBlockchainIconName = tokenIconBlockchainIconName
-        self.isCustomToken = isCustomToken
-    }
-}
-
 struct SendAmountSummaryView: View {
     let viewModel: SendAmountSummaryViewModel
