+++ Tangem/Modules/Send/Summary/SendDestinationSummaryView.swift
@@ -8,21 +8,6 @@
 
 import SwiftUI
 
-class SendDestinationSummaryViewModel: Identifiable {
-    let type: Type
-
-    init(type: Type) {
-        self.type = type
-    }
-}
-
-extension SendDestinationSummaryViewModel {
-    enum `Type` {
-        case address(address: String)
-        case additionalField(type: SendAdditionalFields, value: String)
-    }
-}
-
 struct SendDestinationSummaryView: View {
     private let viewModels: [SendDestinationSummaryViewModel]
